Adds the RestApiEndpoint annotation and cucumber tests
This update includes the addition of the RestApiEndpoint.  This annotation is the main item
a user will interact with in this api.  This update adds in support for cucumber tests which
run on the vintage junit 4 engine from junit 5 platform.  Creates test stubs and enums that
support upcoming testing for many parts of the project.

Resolves: https://github.com/amutsch/jaxrs-autoconfig/issues/3