Adds Scanner capabilities to find RestApiEndpoint annotations
Adds interfaces and a default scanner to find the RestApiEndpoints in the system
and transform them via the Default Transformer that currently exists in the system.
The transformer is autowired in case the user doesn't want to use the default
transformer.

Adds functional tests to verify the functionality based on the default scanner processing
data using the default transformer.

Resolves: https://github.com/amutsch/jaxrs-autoconfig/issues/7